requirements.txt
asgiref==3.8.1
Django==4.2.20
djangorestframework==3.16.0
numpy==2.0.2
opencv-python==4.11.0.86
sqlparse==0.5.3
typing_extensions==4.13.2
pdf2image~=1.17.0